requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.3
djangorestframework==3.16.1
idna==3.11
midtransclient==1.4.2
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
